EDGE-CASES.txt: New file. Also changed incant.sh.

+++ incant.sh
@@ -1,5 +1,15 @@
-#!/usr/bin/env bash
+#!/bin/sh
 
+echo "Beginning Incantation…"
+
+if [ ! -x "./bootstrap" ]; then
+    echo "./bootstrap script not found! Are You in the project directory?";
+    exit;
+fi
+if [ -f "build" ]; then
+    echo "File 'build' needs renamed, so we can have a build/ directory.";
+    exit;
+fi
 rm -rfv build/
 mkdir -pv build/
 ./bootstrap
